+++ flake.nix
@@ -11,20 +11,22 @@
       polysemy-plugin = hackage "0.4.1.0" "117g92l1ppsqd3w0rqjrxfk0lx6yndd54rpymgxljilnv43zg29s";
     };
 
-    ghc921 = { hackage, ... }: {
+    ghc922 = { hackage, ... }: {
       polysemy = hackage "1.7.1.0" "0qwli1kx3hk68hqsgw65mk81bx0djw1wlk17v8ggym7mf3lailyc";
       polysemy-plugin = hackage "0.4.3.0" "1r7j1ffsd6z2q2fgpg78brl2gb0dg8r5ywfiwdrsjd2fxkinjcg1";
       type-errors = hackage "0.2.0.0" "19km8k3g5gn69vvpq0qi0wzsrcjz5k6kb6qjrccq7za39z9slr2z";
     };
 
-  in hix.lib.flake {
+  in hix.lib.flake ({ config, lib, ... }: {
     base = ./.;
-    overrides = { inherit all ghc921; };
+    overrides = { inherit all ghc922; };
     packages = {
       incipit-base = ./packages/incipit-base;
       incipit-core = ./packages/incipit-core;
     };
     main = "incipit-core";
-    hackage.versionFile = "ops/hpack/shared/meta.yaml";
-  };
+    hpack.packages = import ./ops/hpack.nix { inherit config lib; };
+    hackage.versionFile = "ops/version.nix";
+    ghci.extensions = ["NoImplicitPrelude"];
+  });
 }