+++ packages/incipit-core/lib/Prelude.hs
@@ -1,6 +1,0 @@
--- |Exposing this module makes it possible to depend on the package without a mixin.
-module Prelude (
-  module Incipit
-) where
-
-import Incipit