+++ packages/incipit-core/changelog.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Remove module exports for `Prelude` to avoid a Stack bug.
+* Rename `Incipit` modules to `IncipitBase` and `IncipitCore`.
+
 # 0.2.0.0
 
 * Add `unitT`, returning `()` in a `Tactical` environment